Switch to using invoke
